FAC-399:feature:move all to assessmentuserrole
