--- flickit-assessment-core/src/main/java/org/flickit/assessment/core/adapter/in/rest/userassessmentrole/GetAssessmentUserRolesResponseDto.java
@@ -1,8 +1,0 @@
-package org.flickit.assessment.core.adapter.in.rest.userassessmentrole;
-
-import org.flickit.assessment.core.application.port.in.userassessmentrole.GetAssessmentUserRolesUseCase.AssessmentUserRoleItem;
-
-import java.util.List;
-
-public record GetAssessmentUserRolesResponseDto(List<AssessmentUserRoleItem> items) {
-}
